Import in-progress command framework code.
I did not bother to commit all of this code until now because I spent a long time trying different approaches to what I wanted Cloud (the command framework) to do. Some of the code in this commit is still work-in-progress (especially the nightscout subcommands) or needs further documentation, but I wanted to get what I had done so far into version control because it's close to the point where mass-porting of the commands to the new framework can begin.