Add support for command/message update handlers in generic `CommandUser` class.
This allows replies to an author's message to be automatically logged as a command response and can be deleted (or possibly in the future, edited) after the fact. In addition, all `reply` functions now have a `markReply` parameter (defaults to true) which specifies if the reply should be marked. This can be disabled for commands which forcefully delete the author's message on purpose (e.g protecting sensitive data).

Previously this was specific to the `JDACommandUser` class, but I later realised the need for being able to bypass marking a reply. Implementing this into the JDA command user would not have fit well, so I generified it and implemented it into `CommandUser`.