fix(quotes): Detect mentions without nicknames
When mentioning a user in a Discord guild, the format can either be `<@USERID>` if the mentioned user has no nickname in the guild, or `<@!USERID>` if they do have a nickname in the guild. The regex for mentions did not have the `!` as optional, so mentioning someone without a nickname would not be matched by this pattern.