fix(docs): Invite link not adding bot
The second `&scope` parameter seemed to be overwriting the first `&scope`, making the invite only have `applications.commands` scope.

In addition, the permission bit was changed to include `Use Slash Commands`. I'm not sure if this was necessary, but it's what was given in `diabot about` so I figured I'd include it.