fix(graph): remove ephemeral calls, don't await for interaction hook.
the ephemeral state of a deferred response is decided by the ephemeral state when calling `deferReply()`. it cannot be changed later (and has no effect if attempted) so including calls to it is misleading.

the interaction hook is provided by the event, so `await`ing for the object through the `deferReply()` response is unnecessary.