feat(graph): use ephemeral response with graphing disabled message.
in most cases, the database will respond within the 3-second window that Discord provides. since this is an error that doesn't need to be seen by other users, and the fact that we're only relying on a connection that should be stable (the database), i think it's fine to only defer a response when we call an outside source (the user's nightscout instance).