refactor: remove button ID map in favor of a prefix-based ID system
this removes the need to have a map of component IDs to the command classes. instead, the component ID will begin with the main command name and delimit with a colon. a helper function for `String`s have been added to automatically generate an ID in this format.

this also adds a default implementation for the `execute(ButtonInteractionEvent)` function which allows removal of the unimplemented button interaction functions in other classes.