refactor(quote): use await instead of blocking
`.block()` returns a nullable value, while `.awaitSingle()` returns a non-nullable value. Because this is in a coroutine scope anyway (and I plan on revamping all the commands to properly use coroutines later) I'm changing this to await in order to take care of both the blocking issue and the nullable issue.